screenshare small refactor
